ASTON-18 [fix]: styling filters component
