+++ src/components/radio/radio.module.scss
@@ -3,39 +3,43 @@ button {
   all: unset;
 }
 
-.RadioGroupItem {
-  background-color: white;
-  width: 25px;
-  height: 25px;
-  border-radius: 100%;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.5);
-}
-.RadioGroupItem:hover {
-  background-color: #f4f0fe;
-}
-.RadioGroupItem:focus {
-  box-shadow: 0 0 0 2px black;
-}
+.RadioGroup {
+  &Item {
+    background-color: white;
+    width: 25px;
+    height: 25px;
+    border-radius: 100%;
+    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.5);
 
-.RadioGroupIndicator {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100%;
-  position: relative;
-}
-.RadioGroupIndicator::after {
-  content: '';
-  display: block;
-  width: 11px;
-  height: 11px;
-  border-radius: 50%;
-  background-color: #f7b171;
+    &:hover {
+      background-color: #f4f0fe;
+    }
+
+    &:focus {
+      box-shadow: 0 0 0 2px black;
+    }
+  }
+
+  &Indicator {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+    height: 100%;
+    position: relative;
+
+    &::after {
+      content: '';
+      display: block;
+      width: 11px;
+      height: 11px;
+      border-radius: 50%;
+      background-color: #f7b171;
+    }
+  }
 }
 
 .Label {
-  color: white;
   font-size: 15px;
   line-height: 1;
   padding-left: 15px;
