--- src/components/radioContainer/radioContainer.module.scss
@@ -1,10 +1,0 @@
-/* reset */
-button {
-  all: unset;
-}
-
-.RadioGroupRoot {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
