+++ src/styles/_boilerplate.scss
@@ -29,6 +29,12 @@ a:visited {
   color: inherit;
 }
 
+fieldset {
+  margin: 0;
+  padding: 0;
+  border: 0;
+}
+
 body {
   margin: 0;
   padding: 0;
